requirements.txt
fastapi
uvicorn
python-multipart
pillow
docling
msal
requests
pywin32
python-dotenv
